[svn-r512] A partial fix for req#3930 (spawnintrax test failing).
In practice, this fix should cause MPI_Intracomm_merge failures to be an
extremely rare situation.  Furthermore, these failures now blow up with an
assertion failure instead of deadlocking the application.

The real fix for this problem involves PM and PMI changes that are out of scope
at this time.