[svn-r3752] Change the atomics library to use an abstract atomic type, MPIDU_Atomic_t.
This change was needed to deal with the alignment requirements of a platform
like BlueGene/P and for platforms where general loads or stores are not always
atomic.  This commit updates the intel_32_64 and sicortex impls but none of the
others, which are now broken by this change.  They should be updated soon.

Reviewed by buntinas@.