[svn-r4001] Compute sub-communicator context IDs from the parent's context ID.
This change removes the MPIR_Allreduces that were added to MPI_Init time as part
of the SMP collectives work.  It also brings the number of MPI_Allreduces on
each communicator creation back down to 1 instead of 3.  It does this by shaving
2 bits from the context ID to indicate parent, intranode, or internode.

This change also removes one of the context suffix (also called offset) bits
from context IDs, for a final number of 1 bit.

The net change in the context ID space after this change is a reduction by a
factor of 2 (1 bit).

Reviewed by buntinas@.