[svn-r4168] Fix for ticket #485, hopefully the last one.
MPIR_CXX_*COMPLEX_VALUE macros are defined whenever --enable-cxx is passed to
configure, whether or not they are equal to MPI_DATATYPE_NULL.  This caused some
problems in the switch/case statements if the C++ complex datatypes were not
supported.  In particular this was a problem with the Intel C++ compiler on the
breadboard cluster nightly tests.

No reviewer.