[svn-r4618] Fix for sendselfth (tt #611).  Also improved progress debug logging.
We weren't initializing the condition variable, so cond_wait would fail,
and the thread in progress_delay wouldn't ever release the associated
mutex.  Since the mutex wasn't ever released, the other thread that was
yielding was unable to reacquire the mutex and check the completion
counter, resulting in a deadlock.

No reviewer.