[svn-r5066] Use the preprocessor instead of configure for mpe_thread.
This commit refactors the mpe_thread.h and mpe_thread.c files so that
they are no longer created by config.status.  Instead, the preprocessor
is used so that changes to the thread functionality may be made without
re-running config.status every time.  This is step 1 in a broader
clean up and refactoring of the threading code in MPICH2.

This commit is also an important precursor to fixing ticket #566 because
two of the remaining poorly named global symbols live in this mpe_thread
package.

This commit breaks the windows build.

Reviewed by buntinas@.