[svn-r5229] Fix potential deadlock in mtest communicator creation.
The 'done' variable that is used to control exit from the communicator
creation loop was not guaranteed to be consistently set.  Since most
iterations of the loop require participation from all processes in
MPI_COMM_WORLD a deadlock was possible if some of the procs escape the
loop before the other procs.  This fix uses an MPI_Allreduce to ensure
that all processes are actually done, not just the local proc.

No reviewer.