[svn-r5259] Set the is_low_group field when creating an intercomm via MPI_Comm_create.
This corrects a bug that was introduced in r5249.  When the is_low_group
field isn't set deadlocks will typically occur for "full-duplex"
collectives over intercommunicators (such as MPI_Allreduce).

This commit also runs each of the intercomm collective tests with a
higher process count to ensure that more intercommunicator code paths
are exercised.  In this particular case running at higher process counts
is an effective regression test.

No reviewer.