[svn-r5272] Fix a multi-threaded context ID leak.
Under multi-threaded contention the old context ID allocation functions
would leak context IDs as part of the error checking.  This commit
refactors the allocation interfaces to minimize programmer confusion and
eliminates the leak in the previous implementation.  Context ID
allocation error checking is also somewhat improved and a regression
test for this issue is included.

Reviewed by buntinas@.