[svn-r5277] Add MPI_Reduce_scatter block (implements MPI Forum ticket #27).
This commit implements MPI Forum ticket #27, which adds a new collective
function named MPI_Reduce_scatter_block.  This is equivalent to
MPI_Reduce_scatter with a consistent block size argument.

NOTE: this particular implementation of MPI_Reduce_scatter block is
inefficient.  It just performs an MPI_Reduce into a temporary buffer on
rank 0 and then scatters that buffer to the rest of the communicator.

For more history and information about this feature, please see the
following URL:

https://svn.mpi-forum.org/trac/mpi-forum-web/ticket/27

No reviewer.