[svn-r5307] Scalable graph topology implementation (MPI Forum ticket #33).
This commit implements MPI Forum ticket #33, which adds a distributed
graph topology specification API.  This implementation does not use the
topology information at this time, but it is stored and can be queried
as specified by the MPI-2.2 Standard.

For more history and information about this feature, please see the
following URL:

https://svn.mpi-forum.org/trac/mpi-forum-web/ticket/33

No reviewer.