[svn-r5362] Fix alltoall{v,w} zero count/type hangs (ticket #872).
This fixes the cases in MPI_Alltoallv/MPI_Alltoallw where the send
counts are 0 with a non-zero-sized send type and the recv counts are
non-zero with a zero-sized recv type.  Prior to this change MPICH2 would
hang when this situation was encountered.  This commit also includes a
regression test for this behavior.

Thanks to Michael Hofmann for reporting the bug and providing a succinct
test case that illustrates the problem.

No reviewer.