[svn-r5374] Fix subarray test failures introduced in r4099.
The MPIDI_CH3U_Request_load_recv_iov function must fill the iov in the
rreq at offset 0 *and* set rreq->dev.iov_offset to 0 because callers of
this function as an OnDataAvail function don't know what action is
occurring in the handler and can't set the offset to zero themselves.

See r4099 and ticket #462 for some info on the previous test1_dt bug.

Reviewed by buntinas@.