[svn-r5480] Added an initial draft of the hwloc process binding library as an alternative to plpa. The following parts still need to be done:
1. Check when hwloc is usable and when it is not. Ideally the hwloc
configure should deal with this, and we should be able to always
enable it.

2. Initialize appropriate structures with hwloc. We might need to
modify the underlying hydra binding infrastructure a bit, since hwloc
is more powerful than plpa (assuming we want to use the extra
features).

3. This commit directly drops in the released version of hwloc, which
contains a bunch of autogenerated files. The problem with using an svn
version is that hwloc uses libtool, while we don't. We need to either
edit the released version of hwloc to manually remove parts of it, or
start using libtool.