[svn-r5499] Make lock-free reference counting work again.
There were several layers of brokenness that prevented lock-free
reference counting from occurring. Some of this code has been brought
over from the threads development branch, but most of it is new.  This
code still contains many questionable uses of the reference count
variables that need to be investigated and fixed (search for FIXME-MT).

No reviewer.