[svn-r5506] Nemesis SendQ fixes for multithreading.
This change adds debug logging to the SendQ enqueue/dequeue macros.  It
also adds reference counting in those macros.  The reference counting is
needed to avoid a race between deallocation and dequeuing in the case
where an OnDataAvail function completes and frees the request before we
know if we should dequeue it from the queue.

No reviewer.