[svn-r5605] Fix MPI_Comm/MPI_Win errhandler leak.
Prior to this commit we only released references to previously set comm
and win errhandlers in MPI_{Comm,Win}_set_errhandler if the newly set
errhandler was *not* builtin.  This resulted in an errhandler object
leak if a user-defined errhandler was set and then a predefined
errhandler was set afterwards.

Reviewed by buntinas@.