[svn-r5890] Revamp the persistent proxy code; mainly motivated by the requirements of dynamic processes. Attempting to move the persistent launching capability to the bootstrap server completely and remove any information corresponding to it in the upper layers.
The idea is to have a separate daemon that plays the role of
ssh/rsh/slurm on systems where any such daemon is not available. It
just boots a process and forwards stdout, stderr, stdin (no Hydra
specific information). The upper layers will always talk to this
daemon as in their "runtime mode". As a bonus, we also get the ability
for multiple jobs to share the persistent daemons and the ability to
add security/authentication later.

Current status: Runtime launching works correctly, but the bootstrap
interface is quite broken making the persistent launch part
difficult. It expects too much information that is of no use to the
bootstrap servers. We'll leave the code in this commit as a WIP
(persistent launching is in a broken state) and get back to it after
fixing the bootstrap interface. The hydserv persistent daemon is yet
to be added.