[svn-r6022] Don't rely on the PAC_SUBCONFIG_INIT/FINALIZE calls. We don't want to use the INTERNALS_CFLAGS and friends set by the rest of MPICH2.
The problem mainly occurs when the rest of MPICH2 is setting the
internal flags to link to a shared library, and Hydra is not able to
see where the shared library is located. This information
(LD_LIBRARY_PATH) can be passed to the MPI processes by setting the
environment through mpiexec, but the only way to pass this information
to mpiexec and friends is by setting it in the default shell path.

Thanks to Brice Goglin for reporting the error.