[svn-r6270] Add AR_FLAGS and set them to Scr on darwin. (ticket #754)
Previously flags to AR were fixed by simplemake to be "cr".  With this
change this can be controlled by configure/make via the new AR_FLAGS
precious variable.  On darwin using "Scr" shaves ~10% off the normal
build time.

This is a minor, nibbling around the edges fix.  The real fix is to
rework all of the linking in MPICH2 so that parallel builds are possible
and only a single invocation of AR/RANLIB occurs for each library that
we produce.

Reviewed by balaji@.