[svn-r6739] avoid MPI_Info key/val const-correctness warnings in ROMIO
Many of the key (and for MPI_Info_set, value) arguments are string
literals, which elicit warnings about "discards qualifier" because the
MPI_Info interface is not itself const-correct.  This commit adds
ADIOI_Info_ helper macros that cast away the const qualifier and
converts most uses of MPI_Info_ functions to use the new macros.

Reviewed by robl@.