[svn-r7007] bugfix: ch3:sock threading was broken by earlier fine-grained commits
Due to the tangled mess of header include ordering, a convenience #ifdef was
checked too early in another header when it was not yet defined.  This led to a
null implementation of MPIDI_CH3I_PROGRESS_WAKEUP in ch3:sock, even though this
is essential functionality for sock.

No reviewer.