[svn-r7191] 1. Simplify some of the process binding code. Also get rid of some redundant checks in hwloc's binding. These errors are already checked for by the top-level hydra code.
2. Do a depth first search of elements for finding what OS indices we
need to bind to. This is cleaner, but if the user requests for only a
few resources to be used (e.g., asks us not to share a core for
multiple processes), but then gives us more processes than what we can
handle without sharing, we will end up allocating multiple processes
on the same OS index. This part will be fixed when we fix ticket #914.