[svn-r7359] fix group_translate_ranks performance for MPI_COMM_WORLD's group
Prior to this commit, the time that MPI_Group_translate_ranks took was
dependent on the _values_ of the input ranks, the size and pattern of
group2, as well as the number of input ranks.  Now for comm_world's
group (and simple subsets thereof) the translation time only depends on
the number of ranks that are translated.

The MPI_COMM_WORLD use case is essential for tools like Scalasca that
must translate communicator ranks to canonical ranks in MPI_COMM_WORLD.

Reviewed by buntinas@.