[svn-r7684] fix the cb_config_list keyval usage in ROMIO
In the old code it was possible that ROMIO freed the keyval multiple
times, rather than just once.  This was fine in MPICH2, which is robust
in the face of such behavior, but caused problems for ROMIO over Open
MPI.  We now utilize the MPI-2.2 LIFO ordering for attribute destruction
in order to free the keyval exactly once.

Reported by Pascal Deveze and related to ticket #222.

Reviewed by robl@.