[svn-r8288] performance fix for bcast via recursive-doubling scatter/allgather
Changeset r3686 refactored the bcast algorithms into separate functions,
but did so incorrectly.  After r3686, the bcast via
scatter-then-allgather that used the recursive-doubling allgather did
not accurately calculate the amount of data present at each process
after the scatter step.  Instead, it performed the equivalent of an
allgatherv with recvcounts[root]=count and recvcounts[non-root]=0.

This error turns an algorithm with approximate cost:

  lgp.alpha + n.((p-1)/p).beta

into an algorithm with approximate cost:

  lgp.alpha + n.lgp.beta

Reviewed by buntinas@.