[svn-r8374] bugfix: MPI_Get_elements with exotic types (tt#1467)
The "type contents" is used to calculate the number of basic elements,
and for some reason the pointers to the type contents were recalculated
at the upper level instead of calling MPID_Type_access_contents.  This
duplicated logic didn't account for padding between the arrays in some
cases, which led to calculations based off of undefined data.  However
only fairly complicated types ever triggered this bug because it requires
both more than one basic element type and a type constructor with an odd
number of ints in the contents.

This commit includes a new "datatype/get-elements" regression test.

Thanks to Fab Tillier @ Microsoft for reporting the bug.

Reviewed by buntinas@.