[svn-r8394] bugfix: use the correct comm in MPIDU_Sched_start_entry
The code did not originally track per-entry communicators, so it was
correct in the past to use a single comm stored in the schedule.

This fixes most intercommunicator collectives, which were broken because
intercomm collectives commonly use multiple comms (the real comm and the
local comm).  SMP collectives similarly were broken.

Reviewed by buntinas@.