[svn-r8400] intercomm Ibarrier modernization and bugfix
The bug fix was a devilish stack corruption bug that was caused by using
a stack variable as the broadcast buffer.  After MPIR_Ibarrier_inter
returned, the memory for the variable was reused for other purposes,
resulting in erratic corruption and segfaults in the progress engine.
Mallocing the buffer instead resolves this problem.

The modernization includes switching to the MPID_SCHED_BARRIER macro and
dispatching sub-collectives via the coll_fns table instead of statically.

No reviewer.