[svn-r8593] replace buggy bubble sort in MPI_Comm_split with insertion sort (tt#1493)
qsort() is still used by default when available.  The bubble sort
implementation previously used was buggy in that it was unstable.  It
did sort correctly, it just was not stable, which is necessary for correct
comm split behavior.  Insertion sort is always better than bubble sort
and just as simple, so we use it instead.  We should consider just
deleting the alternative sorting algorithm and using qsort exclusively.

Thanks to hjcao@nudt.edu.cn for the bug report.

Reviewed by buntinas@.