[svn-r8612] add conditional `extern "C" {}' guards in mpix.h for C++ compilers
The extensions in mpix.h are all written in C and are exported as C
symbol names.  When mpix.h is included by a C++ compiler, the C++ must
understand that these are C prototypes and not attempt to generate
undefined symbols that point to mangled C++ versions of the symbol
names, otherwise an error will occur when the application is linked.

No reviewer.