[svn-r8702] extend the timeout for the nonblocking{2,3} tests to 7 minutes
The nonblocking3 test, in particular, was failing with timeouts in the nightly
tests.  420 seconds is the value we use for bcast2 and bcast3, so it seems
reasonable to use the same value here.  If nonblocking3 continues to fail with
timeouts with the new value then there is probably a hang occurring in the test
or the MPI library.

No reviewer.