[svn-r9006] pkg-config support improvements.
1. Get rid of unnecessary README which gives no useful information.

2. Combined the pkg-config files for different bindings into a single
file.

3. Bug-fix: When a separate library for the profiling symbols is not
available, we were ending up with a hanging "-l".

4. Bug-fix: pkg-config files do not allow shell code to find the value
of a variable. This logic has to go into configure. No idea how this
was supposed to be working earlier.

5. CXXFLAGS, FFLAGS and FCFLAGS, are not directly supported by
pkg-config, so we don't expose them as first-class variables but still
allow them to be queried using the --variable option.

This patch still doesn't add Requires and Requires.private variables.

Reviewed by buntinas and chan.