[svn-r9144] add a "pre-update hooks" facility to maint/updatefiles
This makes it easier to mimic svn:externals behavior when using git-svn.  It also
permits other local fixups and changes because of a nonstandard environment.  The
intent is that these hooks should not be committed, but are only for local scripts.

Corresponding post-update hooks would be easy to add if we ever have a use case.

No reviewer.