[svn-r9243] convert to PAC_SUBCFG_AUTO_SUFFIX in all subconfigure.m4 files
This was the original plan, but the first implementation of the macro
was a bit buggy, so everything was temporarily hard coded instead.  All
new subconfigure.m4 files should use this macro, since it greatly
reduces the chance of making an error by hand writing the name and makes
the subconfigure.m4 file easier to move to another directory.

No reviewer.