[svn-r9246] use '/' in subconfigure.m4 structured comments for dependencies
Using '_' and '/' interchangeably was bound to lead to problems later on
when someone added a subconfigure.m4 directory with name that contained
an actual '_'.

Downstream projects like MVAPICH2 or Cray that may have migrated netmods
or other subsystems to the new build system will need to update any
structured comments to use the newer format.

No reviewer.