[svn-r9465] warning stomp: fix most uses of "index" & "reindex" as locals/parameters
They shadow the global symbols of the same name and make recent GCC
very unhappy when building with --enable-strict.

I left "index" as the parameter in MPI_Testany/MPI_Waitany for now, since
these names are mandated by the MPI Standard.  But it would be nice to
change them or suppress them in order to get a cleaner build.

No reviewer.