[svn-r9616] tt#1547: fix darwin dylib builds with clang
We weren't linking libpmpich.dylib against libmpich.dylib or any of its
dependencies, so this somehow confused the linker about the TLS-ness of
the MPIR_Thread variable.  I'm a little confused still about why this
wasn't a problem for gcc-driven links, but it doesn't matter that much.

Thanks to Sean Farley for reporting the bug.

No reviewer.