[svn-r9758] use better algorithms in MPIR_Allreduce_group
This is a copy of the non-SMP-aware code in MPIR_Allreduce_intra that
has been modified to perform rank translation immediately prior to any
communication calls.  For cases where incremental group rank translation
is abnormally slow compared to batch translation, this is a non-optimal
method.  A better approach would be to compute the list of ranks that we
communicate with and then translate them all at once.  Such an approach
makes the code much uglier though, so I'm punting on that until we
experience such a case.

No reviewer.