[svn-r9981] tt1645: fix object handle bugs/scalability issues
The handle allocator code had some confusingly named constants that led
to a bug where the two different bit fields were being put in opposite
places.  This wasn't a big problem for small request sizes, since they
were being treated consistently, but the two bit-fields had different
sizes, so when increasing this number one of the fields overflowed and
caused problems.  I've renamed the constants so that they make a bit
more sense and made sure that we are using them consistently.

I also increased the default limit on the number of handles from 256 Ki
to 8 Mi, which should be more appropriate for petascale+ systems.

Thanks to Bob Cernohous @ IBM for the detailed bug report and suggested
fixes.

No reviewer.