[svn-r10106] implement neighborhood collectives
The default nonblocking implementations use the naive "as-if" code from
the draft MPI-3 standard.  No attempt is made to optimize the
communication patterns at this time.

The default blocking implementations simply call the corresponding
nonblocking versions.  NBC overheads are unlikely to be too large for
most neighborhood collective use cases.  If this assertion turns out to
be false at some point then we can revisit this implementation choice.

Reviewed by buntinas@.  (modulo a few last minute fixes)