[svn-r10219] Refactoring of oputil for use in RMA.
An extra argument was added to the oputil macros to provide a name,
which can be used to generate union types.  This is needed for
generating types to be included in the RMA packet headers.  The RMA code
is also modified to use this new functionality, replacing the old
hand-coded unions.

Reviewer: goodell@