[svn-r10247] Get rid of duplicate jump on failure for the MPIR_ERRTEST_ macros.
In several places, after checking for a parameter (e.g., comm) we were
directly using it assuming that the parameter is valid.  Since the
previous ERRTEST macros did not jump to fn_fail on an error, this
could result in undefined behavior if the parameter was invalid.  Now,
since we jump on errors within the macros themselves, once the check
is done, we know that the parameter values are valid.

Reviewed by buntinas.