[svn-r10312] tt#1679: fix SIGBUS on Solaris/SPARC
By converting the scratchpad structs into a union and then adding a pointer
type, the compiler becomes more inclined to align the field on a valid pointer
boundary.  This is a temporary, low-risk, low-ish-robustness fix for the problem
before the 1.5 release.  The proper fix is to stop aliasing the character arrays
with the structures and instead just place those structures in the VC via some
macro hook (think `MPID_REQUEST_DECL`).

Reviewed by buntinas@.