[svn-r10423] Added passive target immediate locking
When enabled, this mode of operation immediately requests the lock when
MPI_Win_lock is called.  Currently, this is enabled by setting the
MPICH_RMA_LOCK_IMMED environment variable.  In the future, we can also make
this mode of operation available though an info/assert.  This capability is
needed to implement MPI-3's flush operations.

Reviewer: buntinas