[svn-r10508] Refactoring RMA Ops list to DL
In this patch, I have refactored the RMA ops list again to use the MPL UTList
doubly-linked list and to treat the list as a proper object.  This should set
us up to work with multiple lists, as we will soon have one list per target.
Doubly-linking the list is a big help in terms of maintainability (no more
prevNext pointers) and flexibility (better implementation of request-based
ops and other optimizations).

Reviewer: goodell