[svn-r10587] RMA epoch tracking
This patch adds code to track the RMA epoch state of the local process.
Currently, we are tracking the synchronization states that are allowed by
MPICH; in the future, we may want to restrict this to only states that are
allowed by the standard.  The addition of epoch tracking has several benefits:

 * It allows us to detect synchronization errors (implemented in this patch).
 * It allows us to implement lock_all more efficiently (implemented in this
   patch).
 * It will allow us to distinguish between active and passive target epochs and
   avoid O(p) op list concatenation (future patch).

Reviewer: balaji