Eliminate enqueueing of lock op in RMA ops list
Prior to this patch, a lock entry was enqueued in the RMA ops list when
Win_lock was called.  This patch adds a new state tracking mechanism, which we
use to record the synchronization state with respect to each RMA target.  This
new mechanism absorbs tracking of lock operation and the lock state at the
target.  It significantly simplifies the RMA synchronization and ops list
processing.

Reviewer: goodell