[svn-r10622] tt#1570: "no" means no for f77/fc/cxx
When --disable-LANG is passed to configure, set the corresponding
language variables to "false" to prevent searching for a working
compiler when the user specifically disabled that language.

If this commit causes any problems, revert it.  We lived without
problems without it for many years, so this isn't a critical feature.

Reviewed by balaji@.