[svn-r10800] ROMIO: check for consistent prealloc sizes
By using MPI_Allreduce to get the maximum value and minimum value of all
sizes, when the two values are identical, all processes have same values
of size.  The problem of checking the sizes with MPI_Bcast is that the
root will pass the check while the others not.

Based on patch 0004 from the second round of IBM's error checking
patches.