[svn-r10820] fix dylib creation on OS X if fortran is disabled
This is a follow-on to r10622 and tt#1570.  Eric Borisch reported that
shared libraries were not generated on Darwin platforms when
`--disable-f77` and `--disable-fortran` are passed.  The fix in r10622
was buggy, but hopefully this one works a bit better.

If you're debugging this in the future, search for `_lt_disable_F77` in
the generated configure file to help understand why we expect `F77=no`
to prevent libtool problems (and possibly help debug the future
problem).

Reviewed by balaji@.