[svn-r10821] tt#1771: allow `+` in `--prefix=`, etc.
Satish pointed out that we were causing problems when the user specified
directories with a `+` in them.  The updated version uses a `|`, which
is a universally poor user choice for directory names (much like `:`
would be).  Single quotes are now also used instead of double quotes.
This sed code is still pretty fragile in the case of exotic path names,
but it's better now.  If this is still a problem for users, the next
step is either to give up on permitting make-time overrides of special
vars like `prefix` or to at least disable that functionality with a
configure switch of some sort.

Reviewed by balaji@.