TAP: add "xfail" support
This is used to indicate tests that are correct, but functionality that
is expected to fail in the current version of the MPI implementation.
This is expressed in `testlist` files with an argument string like
`xfail=ticketXYZ` or `xfail=REASON` after the program name.
Unfortunately, spaces, quotes, and `#` characters are not permitted in
the format at this time, due to the primitive nature of the testfile
parser code.

If strict MPI testing is requested (by configuring the test suite with
`--enable-strictmpi`) then the `xfail=BLAH` arguments are ignored under
the assumption that they are only accurate for the current version of
the MPICH implementation.

No xfail support has been added to the XML output at this time.

Reviewed-by: dinan