Consolidated RMA op finalization code
This patch consolidates the synchronization and tracking of RMA operations into
a single routine that is called whenever we complete an operation.  The only
exception are lock-op-unlock operations that are completed from within the lock
operation processing code.

This code is pretty ugly, but it will get cleaner once packet flags are been
added.

Reviewer: goodell