Allow f90 mod files to be installed in an arbitrary directory.
Instead of using 'includedir' to install the mod file headers the
automake 'modincdir' is used instead. For example:

  ./configure --includedir=$some_dir modincdir=$some_dir/gnu

This is needed because the fortran mod files are essentially
pre-compiled headers and are therefore compiler-specific. This means
that an application can not compile with xlf90 using a mpif90 script
that was configured and installed using gfortran. By separating out the
mod files generated by gfortran and xlf90 we can create compile scripts
that use the appropriate mod files depending on the compiler used.

Note: For some reason the normal configure processing does NOT check for
      a user-specified 'modincdir' variable and, if not set, default to
      use 'includedir'.

(ibm) CPS 92BQ9Q
(ibm) 52931bf950b389d49177a9aa94acf01c3583a8ce