Fix type incompat. for size_t in BSend degug msgs
This patch updates BSend debugging messages to be compatible with new size_t
arguments added in [7fa7e2cf].  Format strings are updated and size_t args are
cast to unsigned long long.  Upconverting to the largest unsinged integer type
is the only safe way to deal with this in pre-C99.  If we were coding to C99,
we could have avoided the type cast and used '%z' directly in the format
strings.  This closes ticket #1786.

Reviewer: goodell