clmake: more patterns and a regex fix
Handle the case where GNU coreutils are installed in some odd `bin`
directory, such as `gnubin`.

Also fix a bad regex pattern (`[^\w]` means match a single character
that is not `\` and is not `w`, rather than the opposite of `\w`, which
is `\W`) by using the `\b` "boundary" atom.

No reviewer.