Remove unnecessary flush for empty epochs
I initially added a conservative flush message for empty epochs (mostly
for documentation purposes).  This is not needed in the current
implementation, since ops are not issued eagerly.  If/when eager ops are
implemented, this patch should be reverted and additional window state
tracking for this case should be added.  In the meantime, I am removing
this code to improve performance.

Reviewer: goodell