Added missing thread safety to Comm_compare
This function was missing thread safety CS_ENTER/EXIT locks.  The global
lock is necessary for this function because it accesses multiple parts
of the comm object, calls several internal routines that assume the
global lock is held, and upates the ref counts on the groups of each
communicator.

Reported by Steve Oyanagi @ Cray.  This closes ticket #1796.

Reviewer: goodell