configure: add `--disable-long-double`
This helps Cray out (and conceivably other vendors), since they ship
several compilers which have different interpretations of what `long
double` should mean (80-bit vs. 128-bit, being most common).

Note that exporting `MPID_NO_LONG_DOUBLE=yes` in the environment before
running `configure` does work, but may not work as intended in the face
of an automake-initiated reconfigure.  Autoconf does not treat this
variable as precious (and we don't ask it to, since it's for
inside-configure-use only), so its value will not be preserved for the
reconfigure.  The same holds for any `MPID_` variables mentioned at the
top of `configure.ac`.

No reviewer.