Fix req. op. completion outside of PT epoch
This is a temporary fix for request-generating operations to allow their
requests to be completed after the user has called unlock on the given
target.  This closes ticket #1801.  Ticket #1741 is still active and is
keeping track of the fact that an implementation of req. ops that tracks
and completes individual operations (rather than the current approach,
which just calls flush) is still needed.

Reviewer: goodell