un-break `--enable-strict` when using clang
It turns out that `--enable-strict` stopped working at some point when
building with clang.  It would treat the conftest program which was
intended to elicit a warning as being completely erroneous.  My fix is
to instead emit a program with more modest intentional bugs so that we
will get simple warnings which most compilers will detect.

Reviewed-by: balaji