Add info hint support to communicators.
Communicators now accept info hints both at creation time
(`MPI_Comm_dup_with_info`) as well as at any later point in their
life cycle (`MPI_Comm_set_info`).

Hints are stored in a hash table, and may be added by defining a
function of type

  int fun(MPID_Comm *, MPID_Info *, void *)

which applies the hint to the communicator, and then registered with
MPICH via the function `MPIR_Comm_register_hint_fn`. The third
(optional) argument can be used to pass extra data to the handler, such
as local context from when it was registered with the runtime.

Reviewed-by: goodell