add `mpiu_uthash.h`
With a couple of MPICH-specific tweaks.

The `mx` netmod uses a non-namespaced "uthash.h" file based on an older
version of the package.  This is probably OK for now, but we run the
risk of symbol conflict and the update is nontrivial since the API
changed between the two versions.

No reviewer.