Corrected MPI-2 C++ interface to include the Distgraph routines
Updated buildiface for C++ (relevant for MPI 2.2) to create the Distgraph
C++ interfaces.  Also fixed errors in the C++ tests where the tests failed
to include the configure-generated configuration file mpitestconf.h before
included files that depended upon that file.

In the process of creating the tests for the C++ Distgraph routines,
discovered errors in the error checking in the MPI dist graph routine to
extract the graph data.  This commit does not fix that error.