tt#1817: move configure subdir exports earlier
We already were exporting these variables for the "main" subdirs, such
as the configured device/channel/netmod.  The bug was that we weren't
setting these exports in time for OPA or MPL to pick them up.  When
[1a1ff11e] came along and reordered the compiler search list, this
caused MPICH to choose `cc` and OPA to choose `gcc`, which don't agree
on the validity of certain warning flags.  So `--enable-strict=c99`
causes OPA's configure to fail because of a bad compiler flag exported
by the top level.

An alternative "fix" to this commit is to ensure that every single
subdir package uses exactly the same logic in order to select compilers
and compiler options.  This seems much more difficult to ensure.

Reviewed-by: balaji